Add some tests for DS APIs, plus a bit of supporting refactoring
